separated checkout mobile tests (#4799)
Closes: #4684
