+++ projects/storefrontapp-e2e-cypress/cypress/integration/mobile/product-search/product-search-product-type-flow-mobile.e2e-spec.ts
@@ -1,5 +1,5 @@
-import * as productSearchTypeFlow from '../../helpers/product-search-product-type-flow';
-import { formats } from '../../sample-data/viewports';
+import * as productSearchTypeFlow from '../../../helpers/product-search-product-type-flow';
+import { formats } from '../../../sample-data/viewports';
 
 context(
   `${formats.mobile.width + 1}p resolution - Product search product type flow`,
