+++ projects/storefrontapp-e2e-cypress/cypress/integration/mobile/product-search/product-search-store-flow-mobile.e2e-spec.ts
@@ -1,6 +1,6 @@
-import * as productSearchStoreFlow from '../../helpers/product-search-store-flow';
-import { clickSearchIcon } from '../../helpers/product-search';
-import { formats } from '../../sample-data/viewports';
+import * as productSearchStoreFlow from '../../../helpers/product-search-store-flow';
+import { clickSearchIcon } from '../../../helpers/product-search';
+import { formats } from '../../../sample-data/viewports';
 
 context(
   `${formats.mobile.width + 1}p resolution - Product search store flow`,
